requirements.txt
requests
pandas
ta
python-dotenv